requirements.txt
# Core dependencies
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0

# Pinecone SDK
pinecone>=5.0.0

# LangChain core + providers you actually use
langchain-core>=0.2.0
langchain-huggingface>=0.3.1
huggingface-hub>=0.23.0

# Gemini via LangChain
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

